be/app/requirements.txt
jaeger-client
Flask-Opentracing
flask
requests